add cdk-monitoring-constructs (#94)
* replace cdk-watchful with cdk-monitoring-constructs

CDK Monitoring Constructs is a logical successor of CDK Watchful. It builds on the original idea, extending it to many more AWS offerings and provides a customization layer. 

You can verify yourself (e.g. by checking my profile or updated Watchful readme), we are now the maintainers of both repos. We plan to keep CDK Watchful in KTLO mode only and focus on the new project instead.

* Update readme.md

added at the end instead of replacing cdk-watchful, since it is still alive